Issue: ERROR:django.request Internal Server Error: /oauth2/oauth2callback/
Explanation:
For some reason, Gapps is not returning a name when it retrieves a user profile.
Profile is expected to have a name key so this breaks the GApps login mechanism.

Solution:
Not sure why GApps is not returning name, seems to be an edge case.
Handle it gracefully, for now. Explore it further if recurring.

Code changes:
Check for keys before processing response from GApps, handle gracefully if not present.
django_gapps_oauth2_login/django_gapps_oauth2_login/utils.py

Update version:
setup.py